Kalman Filter localization running on background.
*Important*: this requires the eBot to be flashed so that it
send all the sensory data (20 values, see eBot.read_all() )
periodically. Currently tested with a period of 50 ms.

Localization (i.e. the result of eBot.position()) is now
computed by the host in a separate thread running on the
background, instead of by the eBot itself.
Exceptions in the thread are not handled, but a possible
way to handle them is commented in eBot.update_background().

A derived class of Serial called SafeSerial is used, where all
reads and writes and flushes are thread-safe by using
a threading.Lock() lock.
By default each instance of SafeSerial creates its own lock,
but the user can specify a lock of his own so eBot-API can
share a lock with other applications that may be accesing the
same critical resources.