Fix connection to eBot; relax timeout
The eBots have a hard time connecting via bluetooth
when there are many of them. To account for this, the
read and write timeout of the serial connection has been
increased to 5 seconds. Also, it will try to connect for
10 times instead of 3.

This may have to be revisited when big numbers of eBot try to
connect at the same time.